requirements.txt
pip>=22.2.2
torch
torchvision
pandas
pytorch-lightning
tensorboard
tabulate
tqdm
pillow
notebook
jupyterlab
matplotlib
seaborn
ipywidgets
boto3
pyprojroot
scikit-learn
opencv-python
huggingface-hub=0.15.1
tokenizers=0.13.3
transformers=4.29.2
datasets=2.12.0
dill=0.3.6
multiprocess=0.70.14
pyarrow=12.0.0
responses=0.18.0
xxhash=3.2.0
